+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unrelease]
+
+### Changed
+
+- The plain text and html versions of the Tika endpoint have been renamed to `as_html` and `as_text`,
+  hopefully to make it clearer about the response type
+
 ## [0.0.2] - 2023-05-31
 
 ### Added