+++ CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unrelease]
+## [Unreleased]
 
 ### Changed
 
 - The plain text and html versions of the Tika endpoint have been renamed to `as_html` and `as_text`,
   hopefully to make it clearer about the response type
+
+### Removed
+
+- The optional dependencies have been removed as Tika does not support HTTP/2 or Brotli
 
 ## [0.0.2] - 2023-05-31